% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batr.R
\docType{package}
\name{batr}
\alias{batr-package}
\alias{batr}
\title{batr: Imports GUANO Annotated Files and Creates Summary Data and Plots.}
\description{
A collection of tools to import and use GUANO metadata in annotated WAV files
of bat recordings. The package depends on comprehensive metadata being
present in the files. The package extracts this data to a text file for
future loading. It calculates summary data and basic plots.
}
\author{
\strong{Maintainer}: Toby Thorne \email{toby@tobythorne.com}

}
