% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{blank_theme}
\alias{blank_theme}
\title{Blank X and Y Axes}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 7.
}
\usage{
blank_theme
}
\description{
Blank X and Y Axes
}
\keyword{datasets}
