% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{blue}
\alias{blue}
\title{The Color Blue}
\format{
An object of class \code{character} of length 1.
}
\usage{
blue
}
\description{
Quick function for "blue". Saves two characters
}
\keyword{datasets}
