% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_GUANO.R
\docType{data}
\name{data.types}
\alias{data.types}
\title{Maps metadata keys to a data type coercion function}
\format{
An object of class \code{list} of length 12.
}
\usage{
data.types
}
\description{
Maps metadata keys to a data type coercion function
}
\keyword{datasets}
