% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{describe}
\alias{describe}
\title{Summarize a Dataframe}
\usage{
describe(
  x,
  na.rm = TRUE,
  interp = FALSE,
  skew = TRUE,
  ranges = TRUE,
  trim = 0.1,
  type = 3,
  check = TRUE,
  fast = NULL,
  quant = NULL,
  IQR = FALSE,
  omit = FALSE,
  data = NULL
)
}
\arguments{
\item{x}{A dataframe}
}
\description{
I found this in some library called "psych".
}
