% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot_helpers.R
\name{.adaptive_date_breaks}
\alias{.adaptive_date_breaks}
\title{Adaptive date breaks and label cleaning helpers}
\usage{
.adaptive_date_breaks(min_date, max_date, date_label = "\%b")
}
\description{
Adaptive date breaks and label cleaning helpers
}
\keyword{internal}
