% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_vet_summary.R
\name{.add_vet_totals_row}
\alias{.add_vet_totals_row}
\title{Add totals row to vetting summary}
\usage{
.add_vet_totals_row(summary_table, stratify_by)
}
\description{
Add totals row to vetting summary
}
\keyword{internal}
