% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_activity.R, R/plot_timing.R
\name{.batr_theme}
\alias{.batr_theme}
\title{Consistent theme for batr plots}
\usage{
.batr_theme(text_size = 10)

.batr_theme(text_size = 10)
}
\description{
Consistent theme for batr plots

Consistent theme for batr plots
}
\keyword{internal}
