% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_site_summary.R
\name{.build_summary_table}
\alias{.build_summary_table}
\title{Build summary table from filtered dataset}
\usage{
.build_summary_table(dataset, species_list, location_list, sort = FALSE)
}
\description{
Build summary table from filtered dataset
}
\keyword{internal}
