% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_activity.R
\name{.calculate_monitoring_period}
\alias{.calculate_monitoring_period}
\title{Calculate monitoring period dates}
\usage{
.calculate_monitoring_period(
  data,
  monitoring_start,
  monitoring_end,
  active_dates
)
}
\description{
Calculate monitoring period dates
}
\keyword{internal}
