% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_vet_summary.R
\name{.calculate_species_vet_stats}
\alias{.calculate_species_vet_stats}
\title{Calculate vetting statistics by species}
\usage{
.calculate_species_vet_stats(dataset)
}
\description{
Calculate vetting statistics by species
}
\keyword{internal}
