% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_vet_summary.R
\name{.check_manually_vetted_field}
\alias{.check_manually_vetted_field}
\title{Check User.Manually.Vetted field exists and has data}
\usage{
.check_manually_vetted_field(dataset)
}
\description{
Check User.Manually.Vetted field exists and has data
}
\keyword{internal}
