% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_vet_summary.R
\name{.check_species_auto_id}
\alias{.check_species_auto_id}
\title{Check for NA in Species.Auto.ID among vetted files}
\usage{
.check_species_auto_id(dataset)
}
\description{
Check for NA in Species.Auto.ID among vetted files
}
\keyword{internal}
