% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reimport_vetted.R
\name{.copy_vetted_files}
\alias{.copy_vetted_files}
\title{Copy vetted files to original locations}
\usage{
.copy_vetted_files(comparison)
}
\description{
Copy vetted files to original locations
}
\keyword{internal}
