% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_site_summary.R
\name{.create_empty_summary_table}
\alias{.create_empty_summary_table}
\title{Create empty summary table with requested structure}
\usage{
.create_empty_summary_table(species_list)
}
\description{
Create empty summary table with requested structure
}
\keyword{internal}
