% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reimport_vetted.R
\name{.discover_files}
\alias{.discover_files}
\title{Discover WAV files in a directory}
\usage{
.discover_files(directory, label, fast_import)
}
\description{
Discover WAV files in a directory
}
\keyword{internal}
