% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reimport_vetted.R
\name{.display_comparison_table}
\alias{.display_comparison_table}
\title{Display comparison table}
\usage{
.display_comparison_table(comparison)
}
\description{
Display comparison table
}
\keyword{internal}
