% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reimport_vetted.R
\name{.display_import_instructions}
\alias{.display_import_instructions}
\title{Display import_guano instructions}
\usage{
.display_import_instructions(WAV_directory, data_path)
}
\description{
Display import_guano instructions
}
\keyword{internal}
