% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_vet_summary.R
\name{.ensure_location_label}
\alias{.ensure_location_label}
\title{Ensure Location_label column exists}
\usage{
.ensure_location_label(dataset)
}
\description{
Ensure Location_label column exists
}
\keyword{internal}
