% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{.filter_summary_dataset}
\alias{.filter_summary_dataset}
\title{Filter dataset by species and location lists}
\usage{
.filter_summary_dataset(dataset, species_list, location_list)
}
\description{
Filter dataset by species and location lists
}
\keyword{internal}
