% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_vet_summary.R
\name{.filter_vetted_dataset}
\alias{.filter_vetted_dataset}
\title{Filter dataset by species and location}
\usage{
.filter_vetted_dataset(dataset, species_list, location_list)
}
\description{
Filter dataset by species and location
}
\keyword{internal}
