% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reimport_vetted.R
\name{.handle_file_issues}
\alias{.handle_file_issues}
\title{Handle file matching issues (missing, conflicts)}
\usage{
.handle_file_issues(match_result, ask_user)
}
\description{
Handle file matching issues (missing, conflicts)
}
\keyword{internal}
