% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_vetting.R
\name{.handle_missing_wavs}
\alias{.handle_missing_wavs}
\title{Handle missing WAV files with user interaction}
\usage{
.handle_missing_wavs(dataset, WAV_directory, interactive)
}
\description{
Handle missing WAV files with user interaction
}
\keyword{internal}
