% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{.load_and_validate_summary_data}
\alias{.load_and_validate_summary_data}
\title{Load and validate data for summary table}
\usage{
.load_and_validate_summary_data(data_path, species_list, location_list)
}
\description{
Load and validate data for summary table
}
\keyword{internal}
