% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_activity.R, R/plot_timing.R
\name{.load_gap_data}
\alias{.load_gap_data}
\title{Load gap data from log files}
\usage{
.load_gap_data(data_path)

.load_gap_data(data_path)
}
\description{
Load gap data from log files

Load gap data from log files
}
\keyword{internal}
