% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_vetting.R
\name{.match_files_to_wavs}
\alias{.match_files_to_wavs}
\title{Match dataset to WAV files}
\usage{
.match_files_to_wavs(dataset, WAV_directory, fast_import)
}
\description{
Match dataset to WAV files
}
\keyword{internal}
