% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reimport_vetted.R
\name{.match_vetted_to_original}
\alias{.match_vetted_to_original}
\title{Match vetted files to original files by filename}
\usage{
.match_vetted_to_original(vetted_files, original_files)
}
\description{
Match vetted files to original files by filename
}
\keyword{internal}
