% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_activity.R
\name{.prepare_activity_data}
\alias{.prepare_activity_data}
\title{Prepare activity count data for plotting}
\usage{
.prepare_activity_data(dataset, species, location_list)
}
\description{
Prepare activity count data for plotting
}
\keyword{internal}
