% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_manual_vetting.R
\name{.prepare_vetting_dataset}
\alias{.prepare_vetting_dataset}
\title{Prepare vetting dataset by loading and filtering}
\usage{
.prepare_vetting_dataset(data_path, no_manual)
}
\description{
Prepare vetting dataset by loading and filtering
}
\keyword{internal}
