% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_vetting.R
\name{.process_simple_sampling}
\alias{.process_simple_sampling}
\title{Process simple (non-stratified) sampling}
\usage{
.process_simple_sampling(dataset, species_list, save_directory, percentage)
}
\description{
Process simple (non-stratified) sampling
}
\keyword{internal}
