% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_vetting.R
\name{.process_stratified_sampling}
\alias{.process_stratified_sampling}
\title{Process stratified sampling by location}
\usage{
.process_stratified_sampling(dataset, species_list, save_directory, percentage)
}
\description{
Process stratified sampling by location
}
\keyword{internal}
