% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_vetting.R
\name{.sample_files}
\alias{.sample_files}
\title{Sample files from dataset based on percentage}
\usage{
.sample_files(dataset, percentage)
}
\description{
Sample files from dataset based on percentage
}
\keyword{internal}
