% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot_helpers.R
\name{.species_label_map}
\alias{.species_label_map}
\title{Build species label mapping for plots}
\usage{
.species_label_map(species_codes, species_names = NULL)
}
\description{
Build species label mapping for plots
}
\keyword{internal}
