% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot_helpers.R
\name{.species_labeller}
\alias{.species_labeller}
\title{Create a labeller for species facets}
\usage{
.species_labeller(species_codes, species_names = NULL)
}
\description{
Create a labeller for species facets
}
\keyword{internal}
