% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_vetting.R
\name{.validate_directories}
\alias{.validate_directories}
\title{Validate directories exist}
\usage{
.validate_directories(WAV_directory, save_directory)
}
\description{
Validate directories exist
}
\keyword{internal}
