% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reimport_vetted.R
\name{.validate_reimport_directories}
\alias{.validate_reimport_directories}
\title{Validate directories for reimport}
\usage{
.validate_reimport_directories(vetted_directory, WAV_directory)
}
\description{
Validate directories for reimport
}
\keyword{internal}
