% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_vet_summary.R
\name{.validate_vet_summary_data}
\alias{.validate_vet_summary_data}
\title{Validate data for vetting summary}
\usage{
.validate_vet_summary_data(data_path, species_list, location_list, stratify_by)
}
\description{
Validate data for vetting summary
}
\keyword{internal}
