% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_activity_timing_plot.R
\name{first_observations_plot}
\alias{first_observations_plot}
\title{Plot Timing of First Observations Relative to Sunset}
\usage{
first_observations_plot(
  data_path,
  species,
  timezone,
  monitoring_start = NULL,
  monitoring_end = NULL,
  gaps = NULL,
  width = 15.9,
  height = 8.43,
  text_size = 10,
  date_label = "\%b",
  save_directory = NULL
)
}
\arguments{
\item{data_path}{Character. Path to an existing RData file.}

\item{species}{Charater. Code for a single species in the datset to use for
graph generation (e.g. "Epfu").}

\item{timezone}{Character. Specify a timezone for calculating sunset times,
e.g. "America/Toronto"}

\item{monitoring_start}{Character, the date the monitoring began (e.g.
"2019-01-01"), leave as NULL to use the earliest data point as the monitoring
start.}

\item{monitoring_end}{Character, the date the monitoring ceased (e.g.
"2019-01-01"), leave as NULL to use the latest data point as the monitoring
end}

\item{gaps}{Object: a \code{gaps_projectname} data frame generated by the
\code{log_file_parser} function. Defaults to \code{null}.}

\item{width}{Number. The width in cm of the plot if saved to file. Default is
25 cm.}

\item{height}{Number. The height of the plot if saved to file. Default is 20
cm.}

\item{text_size}{Numeric: adjusts the size of text in the figure.}

\item{date_label}{Date value: adjusts the formatting of the month labels on
the y-axis. See https://www.statmethods.net/input/dates.html for formatting.}

\item{save_directory}{Character: if provided a .png image of the plot will be
saved in the folder specified. Defaults to \code{NULL}: no output saved.}
}
\value{
A plot as an object in the current envrionment, and a saved image if
 selected.
}
\description{
\code{first_observations_plot} creates a plot showing the first observation of
the specified species relative to local sunset for each site withn a project.
}
\seealso{
Other Activity Timing: 
\code{\link{old_first_observations_plot}()}
}
\concept{Activity Timing}
