% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gen.mat.to.edge.list}
\alias{gen.mat.to.edge.list}
\title{Edge list from Matrix}
\usage{
gen.mat.to.edge.list(mat, symmetric = TRUE, diagonal = FALSE, text = FALSE)
}
\description{
Creates an edge list from an adjacency matrix.
}
