% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{\%!in\%}
\alias{\%!in\%}
\title{Negation}
\usage{
... \%!in\% NA
}
\description{
A "not in" operator for working with lists.
}
\examples{
bad <- c("this", "that", "other_thing")
good <- c("this", "cat", "that", "dog", "other_thing", "bat")
good \%!in\% bad

}
