% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Importing.R
\name{GUANO_loader}
\alias{GUANO_loader}
\title{Load Previously Read GUANO Metadata From Text File}
\usage{
GUANO_loader(file, project_name, location = "SB.Town")
}
\arguments{
\item{file}{Character, a \strong{full path} to a text file created by
GUANO.reader.}

\item{dataset_name}{Character, a brief and relevant reference phrase that will
be appended to the function outputs for identification.}

\item{town.locaion}{Numeric. Selects the format used to record location in
file metadata. \strong{1:} (default) location recorded in 'SB.Town' field.
\strong{2:} location recored in 'Town' field - legacy from older versions of
SonoBat Data Wizard. \strong{3:} location recorded in 'SB.Location' -
required for some TRCA data.}
}
\value{
A series of dataframes summarising metadata from the files and
 facilitating further analyses and plots
}
\description{
\code{GUANO_loader} imports a text file of GUANO metadata created by
\code{GUANO-reader} and created a number of dataframes to provide summaries
and facilitate further analyses.
}
\examples{
\dontrun{
GUANO.loader("C:/Folder/Folder/Project_Name.txt", "Project_Name")
}
}
\seealso{
Other Import Functions: 
\code{\link{GUANO_reader}()},
\code{\link{import_GUANO}()},
\code{\link{log_file_parser}()}
}
\concept{Import Functions}
