% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_importing.R
\name{GUANO_reader}
\alias{GUANO_reader}
\title{Read GUANO Metadata From Files}
\usage{
GUANO_reader(folderpath, project_name)
}
\arguments{
\item{folderpath}{Character, a path to the folder containing the files you
wish to read the metadata from.}

\item{project_name}{Character, a brief and relevant reference phrase that will be used
to name the text file.}
}
\value{
A textfile in the parent folder of the files folder.
}
\description{
\code{GUANO_reader} reads available GUANO metadata embedded in a folder of bat
acoustic. wav files processed by SonoBat. It creates a text file output of
this metadata in the parent folder of the files that can be loaded for further
analyses by \code{GUANO_loader}.
}
\section{Note}{
 This function will take a long time for large datasets. Folders
 containing tens of thousands of files may take several hours to read!
 However, once complete this action does not need to repeated unless any
 changes are made to the original files. For this reason it is recommended to
 complete all manual vetting before proceeding.
}

\examples{
\dontrun{
GUANO.reader("C:/Folder/Folder/File_Folder", "Project_NA")
}
}
\seealso{
Other Import Functions: 
\code{\link{GUANO_loader}()},
\code{\link{log_file_parser}()}
}
\concept{Import Functions}
