% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_Manual_Vetting_Tools.R
\name{high_frequency_extractor}
\alias{high_frequency_extractor}
\title{Copy or Move a Random Subset of High Frequency Files to a New Sub Directory for Manual Vetting}
\usage{
high_frequency_extractor(
  sonobat_output,
  directory,
  percentage = 0.1,
  move_files = TRUE
)
}
\arguments{
\item{sonobat_output}{Character, a \strong{full path} to a text file output of
the directory created from SonoBat > SonoVet > Export > Use Current Layout}

\item{directory}{Character, a \strong{full path} to the folder containg the
files you wish to subset.}

\item{percentage}{Number. What proportion of files do you wish to extract.
Defaults to 0.1 - 10 percent.}

\item{move_files}{Logical vector, defaults to \code{TRUE}. If \code{TRUE} the
function will \strong{move} selected files into a new directory. If
\code{FALSE} the function will \strong{copy} files into a new directory.}
}
\value{
A folder within the file folder filled with a percentage of high
 frequency files moved or copied.
}
\description{
\code{high_frequency_extractor} moves or copies a proportion of high frequency
files from a directory. To do so reqruies a text file output to be generated
for the directory in SonoBat.
}
\section{Note}{
 this function will only work for files within a single
 directory. \strong{It will not work correctly where files are arranged in
 sub-directories!}
}

\examples{
\dontrun{
high_frequency_extractor("C:/Folder/Files_Folder/sonobat_output.txt", "C:/Folder/Files_Folder", percentage = 0.1, move_files = TRUE)
}
}
\seealso{
Other Manual Vetting Functions: 
\code{\link{manual_vet_extractor}()},
\code{\link{manual_vet_extractor_old}()}
}
\concept{Manual Vetting Functions}
