% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{howmuch}
\alias{howmuch}
\title{Count Unique Values}
\usage{
howmuch(x)
}
\arguments{
\item{x}{A vector like object such as a list or data frame column}
}
\description{
Counts the unique values in a vector or vector like object
such as a data frame.
}
