% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_species_setter.R
\name{list_all_species}
\alias{list_all_species}
\title{Returns a data frame listing all unique species identifications included in a
dataset in an RData file}
\usage{
list_all_species(data_path)
}
\arguments{
\item{data_path}{Character. Path to an RData file to add activity data to.
Optional, a save location will be created before data are saved.}
}
\value{
A data frame containing all unique species identifications.
}
\description{
Returns a data frame listing all unique species identifications included in a
dataset in an RData file
}
\examples{
\dontrun{
set_species_list("C:/Folder/Folder/Data.RData")
}
}
\seealso{
Other species tools: 
\code{\link{set_species_list}()},
\code{\link{show_species_list}()}
}
\concept{species tools}
