% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_reports.R
\name{list_report_templates}
\alias{list_report_templates}
\title{List available report templates}
\usage{
list_report_templates()
}
\value{
Character vector of template names
}
\description{
Returns a character vector of available batr report templates that can be
used with \code{\link{use_report_template}}.
}
\examples{
list_report_templates()
}
