% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_Data_Importing.R
\name{log_file_parser}
\alias{log_file_parser}
\title{Import and Interpret Wildlife Acoustics and Anabat Log Files}
\usage{
log_file_parser(path, dataset_name, monitoring_start, monitoring_end)
}
\arguments{
\item{path}{Character, a path to a parent directory containing all the
relevant log files, following the filename convention described in Note.}

\item{dataset_name}{Character, a brief and relevant reference phrase that will
be used to name the text file.}

\item{monitoring_start}{Character, the date the monitoring began (e.g.
"2019-01-01")}

\item{monitoring_end}{Character, the date the monitoring ceased (e.g.
"2019-01-01")}
}
\value{
A series of dataframes summarising activity and uptimes from the files
 and facilitating further analyses and plots
}
\description{
\code{log_file_parser} automatically ingests the logfiles created by Wildife
Acoustics SM3 and SM4, and Anabat Swift acoustic monitoring units. It
identifies, by site, whether a recorder was active on each night in the
monitoring period, creates a data frame of 'gaps' for use in plots, and
calcualtes uptime for each monitor.
}
\section{Note}{
 To work as expected, log files must be renamed following the
 conventions outlined below!

 \strong{Wildlife Acoustics:} These files should be renamed so that each
 filename begins with the site name followed by a hyphen and any additional
 information (e.g. date) that is desired in the file name. For example:
 "site_1-2019.txt". The hyphen and all characters that follow it will be
 disregarded. The site name should match the site name assigned in GUANO
 metadata so that the log file data is appropriately matched with the
 acoustic data.

 \strong{Anabat Swift:} These fiels do not need to be renamed, but files for
 each site should be placed in their own sub-folder. The folder name should
 match the site name assinged in the GUANO metadata of the corrisponding .wav
 files so they can be matched.
}

\examples{
\dontrun{
log_file_parser("C:/Folder/Folder/Log_Files_Folder", "Project_Name", "2019-01-01", "2019-12-31")
}
}
\seealso{
Other Import Functions: 
\code{\link{GUANO_loader}()},
\code{\link{GUANO_reader}()},
\code{\link{import_guano}()}
}
\concept{Import Functions}
