% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manual_Vetting_Tools.R
\name{manual_vet_extractor}
\alias{manual_vet_extractor}
\title{Copy a Random Subset of Files to a New Directory for Manual Vetting}
\usage{
manual_vet_extractor(
  dataset,
  directory,
  species_list = c("Epfu", "Labo", "Laci", "Lano"),
  percentage = 0.05,
  no_manual = FALSE
)
}
\arguments{
\item{dataset}{Character, a "raw_data_..." output from \code{GUANO_loader}.}

\item{directory}{Character, the directory where selected files shoud be copied
to.}

\item{species_list}{Character, a single species or list in the form of four
character species codes used by SonoBat, e.g. "Epfu". Defaults to non
Ontario SAR: \code{c("Epfu", "Labo", "Laci", "Lano")}.}

\item{percentage}{Number. What proportion of files do you wish to extract.
Defaults to 0.05 - 5 percent.}

\item{no_manual}{Logical vector, defaults to \code{FALSE}. If \code{TRUE} the
function will ignore files that have already been assigned an ID.}
}
\value{
A folder within the file folder filled with a percentage of files of
 the selected copied.
}
\description{
\code{manual_vet_extractor} copies a randomly selected subset of files in a
project to a new folder to faciiliate manual vetting. The function will only
move files with an accepted automatic species assingment from SonoBat.
}
\examples{
\dontrun{
manual_vet_extractor("raw_data_project", "C:/Folder/Folder/File_Folder")
}
}
\seealso{
Other Manual Vetting Functions: 
\code{\link{high_frequency_extractor}()}
}
\concept{Manual Vetting Functions}
