% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{matrix2edges}
\alias{matrix2edges}
\title{Edge list from Matrix}
\usage{
matrix2edges(mat, symmetric = TRUE, diagonal = FALSE, text = FALSE)
}
\description{
Creates an edge list from an adjacency matrix.
}
