% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{monitoring_effort_plot}
\alias{monitoring_effort_plot}
\title{Plot of Monitoring Effort by Site}
\usage{
monitoring_effort_plot(
  gaps,
  project_name,
  monitoring_start,
  monitoring_end,
  survey_year = NULL,
  save_directory = NULL,
  title = TRUE,
  width = 25,
  height = 20,
  text_size = 8,
  date_label = "\%b"
)
}
\arguments{
\item{gaps}{Object: a \code{gaps_projectname} data frame genereted by the
\code{log_file_parser} function. Defaults to \code{null}.}

\item{project_name}{Character, a brief and relevant reference phrase that will
be used to name the text file. Ideally mathcing project name used in other
functions.}

\item{monitoring_start}{Character, the date the monitoring began (e.g.
"2019-01-01")}

\item{monitoring_end}{Character, the date the monitoring ceased (e.g.
"2019-01-01")}

\item{survey_year}{Character: for projects with data from multiple years, to
select year}

\item{title}{Logical vector. Defaults \code{TRUE}, if \code{FALSE} title is
omitted}

\item{width}{Number. The width in cm of the plot if saved to file. Default is
25 cm.}

\item{height}{Number. The height of the plot if saved to file. Default is 20
cm.}

\item{text_size}{Numeric: adjusts the size of text in the figure.}

\item{save.directory}{Character: if provided a .png image of the plot will be
saved in the folder specified. Defaults to \code{NULL}: no output saved.}

\item{date_breaks}{Date value: adjusts the formatting of the month labels on
the y-axis. See https://www.statmethods.net/input/dates.html for formatting.}
}
\value{
A plot as an object in the current envrionment, and a saved image if
 selected.
}
\description{
\code{monitoring_effort_plot} creates a plot showing monitor uptime at each
site within the project.
}
\examples{
\dontrun{
species_daily_site_plot(species_night_site_projectname, "Project Name", "Mylu", "2019-01-01", "2019-12-31")
}
}
\seealso{
Other Basic Plots: 
\code{\link{multi_species_daily_site_plot}()},
\code{\link{species_daily_site_plot}()},
\code{\link{species_site_aggregated_plot}()}
}
\concept{Basic Plots}
