% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_activity.R
\name{monthly_activity_plot}
\alias{monthly_activity_plot}
\title{Monthly Activity Plot}
\usage{
monthly_activity_plot(
  species_night_site,
  monthly_active_nights,
  exclude_species = NULL,
  species_codes = c("Epfu", "Labo", "Laci", "Lano", "Myle", "Mylu", "Myse", "Mysp",
    "Mysp_all", "Pesu"),
  species_names = c("Big Brown Bat", "Eastern Red Bat", "Hoary Bat", "Silver-haired Bat",
    "Eastern Small-footed Myotis", "Little Brown Myotis", "Northern Myotis",
    "Unidentified Myotis", "Combined Myotis", "Tri-colored Bat")
)
}
\arguments{
\item{species_night_site}{Data frame with Night, Location, Species, Count columns.}

\item{monthly_active_nights}{Data frame with monthly survey effort.}

\item{exclude_species}{Character vector of species codes to exclude.}

\item{species_codes}{Character vector of species codes.}

\item{species_names}{Character vector of species names matching codes.}
}
\value{
A ggplot2 plot object.
}
\description{
\code{monthly_activity_plot} creates a plot showing average nightly activity
by month and species.
}
\seealso{
Other Activity Plots: 
\code{\link{monitoring_effort_plot}()},
\code{\link{species_daily_site_plot}()}
}
\concept{Activity Plots}
