% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{multi_species_daily_site_plot}
\alias{multi_species_daily_site_plot}
\title{Plot Activity for All Species at Each Site Within a Project}
\usage{
multi_species_daily_site_plot(
  dataset,
  project_name,
  monitoring_start,
  monitoring_end,
  gaps = NULL,
  survey_year = NULL,
  save_directory = NULL
)
}
\arguments{
\item{dataset}{Object: a \code{species_night_site_projectname} data frame
genereted by the \code{GUANO_loader} function.}

\item{project_name}{Character, a brief and relevant reference phrase that will
be used to name the text file. Ideally mathcing project name used in other
functions.}

\item{monitoring_start}{Character, the date the monitoring began (e.g.
"2019-01-01")}

\item{monitoring_end}{Character, the date the monitoring ceased (e.g.
"2019-01-01")}

\item{gaps}{Object: a \code{gaps_projectname} data frame genereted by the
\code{log_file_parser} function. Defaults to \code{null}.}

\item{survey_year}{Character: for projects with}

\item{save.directory}{Charactor: if provided a .png image of the plot will be
saved in the folder specified. Defaults to \code{NULL}: no output saved.}

\item{text_size}{Numeric: adjusts the size of text in the figure.}

\item{date_breaks}{Date value: adjusts the formatting of the month labels on
the y-axis. See https://www.statmethods.net/input/dates.html for formatting.}
}
\value{
A plot as an object in the current envrionment, and a saved image if
 selected.
}
\description{
\code{multi_species_daily_site_plot} is a wrapper to apply \code{species_daily_site_plot} creates a series plot of nightly activity for a chosen
species at all sites within a project.
}
\examples{
\dontrun{
species_daily_site_plot(species_night_site_projectname, "Project Name", "Mylu", "2019-01-01", "2019-12-31")
}
}
\seealso{
Other Basic Plots: 
\code{\link{monitoring_effort_plot}()},
\code{\link{species_daily_site_plot}()},
\code{\link{species_site_aggregated_plot}()}
}
\concept{Basic Plots}
