% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{net2df}
\alias{net2df}
\title{Calculate Network Statistics and Pack into DataFrame}
\usage{
net2df(g)
}
\arguments{
\item{g}{An igraph network object}
}
\value{
Dataframe with network statistics from network.
}
\description{
Calculates a number of network statistics and packs them
all into a dataframe.
}
\examples{
df <- net2df(g)

}
