% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{red}
\alias{red}
\title{The Color Red}
\format{
An object of class \code{character} of length 1.
}
\usage{
red
}
\description{
Quick function for "red". Saves two characters
}
\keyword{datasets}
