% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{round_df}
\alias{round_df}
\title{Round Dataframe Values}
\usage{
round_df(x, digits)
}
\arguments{
\item{x}{A dataframe}

\item{digits}{The number of digits to round to}
}
\value{
A dataframe with numeric columns rounded
}
\description{
Rounds the numeric columns in a data frame to the
specified digit. Good for outputting tables in an
RMarkdown document.
}
\examples{
df <- round_df(df)

}
