% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{scatter_theme}
\alias{scatter_theme}
\title{Scatterplot Theme}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 5.
}
\usage{
scatter_theme
}
\description{
A minimal Scatter plot theme for ggplot and sets font to Roboto
}
\keyword{datasets}
