% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_locations_setter.R
\name{set_location_list}
\alias{set_location_list}
\title{Save a Location List into an RData File}
\usage{
set_location_list(location_list, data_path)
}
\arguments{
\item{data_path}{Character. Path to an RData file to add activity data to.
Optional, a save location will be created before data are saved.}

\item{species_list}{A vector list of species codes that correspond to those
in the GUANO metadata for the files in the data set}
}
\value{
An RData file containing a list of species for the project, along
  with other data if added to an existing RData file
}
\description{
Saves a location list into a batr RData file so that you don't have to specify it
each time you run a function. If no location list is specified functions will ask
for it.
}
\examples{
\dontrun{
set_location_list(c("location1", "location2"), "C:/Folder/Folder/Data.RData")
}
}
\seealso{
Other location tools: 
\code{\link{list_all_locations}()},
\code{\link{show_location_list}()}
}
\concept{location tools}
