% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_species_setter.R
\name{set_species_list}
\alias{set_species_list}
\title{Save a Species List into an RData File}
\usage{
set_species_list(species_list, data_path)
}
\arguments{
\item{species_list}{A vector list of species codes that correspond to those
in the GUANO metadata for the files in the data set}

\item{data_path}{Character. Path to an RData file to add activity data to.
Optional, a save location will be created before data are saved.}
}
\value{
An RData file containing a list of species for the project, along
  with other data if added to an existing RData file
}
\description{
Saves a species list into a batr RData file so that you don't have to specify
it each time you run a function. If no species list is specified functions
will ask for it.
}
\examples{
\dontrun{
set_species_list(c("Epfu", "Labo"), "C:/Folder/Folder/Data.RData")
}
}
\seealso{
Other species tools: 
\code{\link{list_all_species}()},
\code{\link{show_species_list}()}
}
\concept{species tools}
