% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_locations_setter.R
\name{show_location_list}
\alias{show_location_list}
\title{Show a Species List that has been saved into an RData File}
\usage{
show_location_list(data_path)
}
\arguments{
\item{data_path}{Character. Path to an RData file to add activity data to.
Optional, a save location will be created before data are saved.}
}
\value{
A vector of the saved Species List, if one exists.
}
\description{
Reveals a Species List that has been saved into an existing Rdata file.
}
\examples{
\dontrun{set_species_list("C:/Folder/Folder/Data.RData")}
}
\seealso{
Other location tools: 
\code{\link{list_all_locations}()},
\code{\link{set_location_list}()}
}
\concept{location tools}
