% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_Plots.R
\name{species_site_aggregated_plot}
\alias{species_site_aggregated_plot}
\title{Plot Species Activity Data for a Project}
\usage{
species_site_aggregated_plot(
  dataset,
  project_name,
  monitoring_start,
  monitoring_end,
  gaps = NULL,
  title = "TRUE",
  save_directory = NULL,
  text_size = 8,
  date_label = "\%b"
)
}
\arguments{
\item{dataset}{Object: a \code{species_night_site_projectname} data frame
generated by the \code{GUANO_loader} function.}

\item{project_name}{Character, a brief and relevant reference phrase that will
be used to name the text file. Ideally matching project name used in other
functions.}

\item{monitoring_start}{Character, the date the monitoring began (e.g.
"2019-01-01")}

\item{monitoring_end}{Character, the date the monitoring ceased (e.g.
"2019-01-01")}

\item{gaps}{Object: a \code{gaps_projectname} data frame generated by the
\code{log_file_parser} function. Defaults to \code{null}.}

\item{title}{Logical vector. Defaults \code{TRUE}, if \code{FALSE} title is
omitted}

\item{text_size}{Numeric: adjusts the size of text in the figure.}

\item{save.directory}{Character: if provided a .png image of the plot will be
saved in the folder specified. Defaults to \code{NULL}: no output saved.}

\item{date_breaks}{Date value: adjusts the formatting of the month labels on
the y-axis. See https://www.statmethods.net/input/dates.html for formatting.}
}
\value{
A plot as an object in the current environment, and a saved image if
 selected.
}
\description{
\code{species_site_aggregated_plot} creates a plot of nightly activity for
each species in a dataset. Different sites within the dataset are aggregated.
Monitor activity can be plotted by including a gaps dataset (generated by
\code{log_file_parser}if desired.
}
\section{Note}{
 Because all sites are aggregated, if gaps are plotted these
 will be applied to each facet of the plot. Overlapping plots are indicated
 by denser colour.
}

\examples{
\dontrun{
species_site_aggregated_plot(species_night_site_projectname, "Project Name", "2019-01-01", "2019-12-31")
}
}
\seealso{
Other Basic Plots: 
\code{\link{monitoring_effort_plot}()},
\code{\link{monthly_activity_plot}()},
\code{\link{multi_species_daily_site_plot}()},
\code{\link{species_daily_site_plot}()}
}
\concept{Basic Plots}
