% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_site_summary.R
\name{summary_table}
\alias{summary_table}
\title{Species Site Summary Table}
\usage{
summary_table(data_path, all_species = FALSE)
}
\arguments{
\item{data_path}{}

\item{all_species}{Logical vector. Default is \code{false}, if \code{true}
creates empty columns for species in species_list that are not present in
the dataset.}
}
\value{
A data frame of observations summarized by location and species.
}
\description{
Returns a table summarizing the total number of observations by site and
location.
}
\examples{
\dontrun{summary_table("C:/Folder/Folder/Data.RData")}
}
