% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{te_pal}
\alias{te_pal}
\title{My Own Palette}
\format{
An object of class \code{character} of length 5.
}
\usage{
te_pal
}
\description{
A random palette of colors that I put together that are all sort of blue.
Probably not good for much of anything.
}
\keyword{datasets}
