% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{te_theme}
\alias{te_theme}
\title{General Purpose Plot Theme}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 5.
}
\usage{
te_theme
}
\description{
Use for plots in general to make them look better
}
\keyword{datasets}
