% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{twenty_twenty_three_pal}
\alias{twenty_twenty_three_pal}
\title{2023 Palette}
\format{
An object of class \code{character} of length 5.
}
\usage{
twenty_twenty_three_pal
}
\description{
A palette of colors for 2023, includes "skyblue", "coral", "seafoam green",
"mauve", and a slate grey. Good for visualizations.
}
\keyword{datasets}
