% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{uc_primary_pal}
\alias{uc_primary_pal}
\title{University of Chicago Primary Colors}
\format{
An object of class \code{character} of length 3.
}
\usage{
uc_primary_pal
}
\description{
The Maroon and Grey of the University of Chicago
}
\keyword{datasets}
