% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{uc_second_pal}
\alias{uc_second_pal}
\title{University of Chicago Secondary Colors}
\format{
An object of class \code{character} of length 7.
}
\usage{
uc_second_pal
}
\description{
Secondary palette from the graphic identity guidelines. Use this
for all sorts of things, including visualizations.
}
\keyword{datasets}
