% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{uc_shade_pal}
\alias{uc_shade_pal}
\title{University of Chicago Shaded Colors}
\format{
An object of class \code{character} of length 6.
}
\usage{
uc_shade_pal
}
\description{
Shaded palette from the graphic identity guidelines. Use this for some darker
colors from the secondary palette
}
\keyword{datasets}
