% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{uc_tinted_pal}
\alias{uc_tinted_pal}
\title{University of Chicago Tinted Colors}
\format{
An object of class \code{character} of length 7.
}
\usage{
uc_tinted_pal
}
\description{
Tinted palette from the graphic identity guidelines. Use this for slightly
different colors from the secondary palette.
}
\keyword{datasets}
