% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{uc_violate_pal}
\alias{uc_violate_pal}
\title{University of Chicago Secondary Colors}
\format{
An object of class \code{character} of length 2.
}
\usage{
uc_violate_pal
}
\description{
Violate palette from the graphic identity guidelines. use this for
highlighting bad things.
}
\keyword{datasets}
