% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batr.R
\docType{package}
\name{batr}
\alias{batr}
\title{batr: Imports GUANO Anotated Files and Creates Summary Data and Plots.}
\description{
A collection of tools to import and use GUANO metadata in annotated .wav
files of bat recordings. The package depends on comphrensive metadata being
present in the files. The package extracts this data to a text file for
future loading. It calculates summary data and basic plots.
}
\section{Foo functions}{
 The foo functions ...
}

