% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_report_tools.R
\name{comsci_report}
\alias{comsci_report}
\title{Create Com. Sci.Report(s)}
\usage{
comsci_report(data_path, save_path, sites, author, species)
}
\arguments{
\item{RData}{file}
}
\value{
A PDF report.
}
\description{
Create Com. Sci.Report(s)
}
\examples{
\dontrun{
Insert Example!
}
}
