% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_activity.R
\name{.prepare_activity_data}
\alias{.prepare_activity_data}
\title{Load gap data from log files}
\usage{
.prepare_activity_data(dataset, species, location_list)
}
\description{
Load gap data from log files
}
\keyword{Prepare}
\keyword{activity}
\keyword{count}
\keyword{data}
\keyword{for}
\keyword{internal}
\keyword{plotting}
