% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_tools.R
\name{.quick_summary}
\alias{.quick_summary}
\title{Create quick summary table}
\usage{
.quick_summary(observations, active_dates = NULL)
}
\arguments{
\item{observations}{Observations data frame.}

\item{active_dates}{Optional active_dates data frame}
}
\value{
A quick summary table of total species observations and survey effort.
}
\description{
Create quick summary table
}
