% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_guano.R
\name{.standardize_swift_locations}
\alias{.standardize_swift_locations}
\title{Standardize GPS Coordinates Per Location}
\usage{
.standardize_swift_locations(observations)
}
\arguments{
\item{observations}{Data.frame containing processed observations.}
}
\value{
Data.frame with standardized coordinates per location.
}
\description{
Some recorders report slightly different GPS coordinates for files from the
same physical location. This function standardizes all coordinates for a given
location to the first recorded value.
}
\keyword{internal}
