% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Activity_Timing.R
\name{first_observations_plot}
\alias{first_observations_plot}
\title{Plot Timing of First Observations Relative to Sunset}
\usage{
first_observations_plot(
  dataset,
  project_name,
  species,
  monitoring_start,
  monitoring_end,
  gaps = NULL,
  location = NULL,
  survey_year = NULL,
  timezone = "EST",
  text_size = 8,
  date_label = "\%b",
  save_directory = NULL
)
}
\arguments{
\item{dataset}{Object: a \code{raw_data_projectname} data frame genereted by
the \code{GUANO_loader} function.}

\item{project_name}{Character, a brief and relevant reference phrase that will
be used to name the text file. Ideally mathcing project name used in other
functions.}

\item{monitoring_start}{Character, the date the monitoring began (e.g.
"2019-01-01")}

\item{monitoring_end}{Character, the date the monitoring ceased (e.g.
"2019-01-01")}

\item{gaps}{Object: a \code{gaps_projectname} data frame genereted by the
\code{log_file_parser} function. Defaults to \code{null}.}

\item{location}{Specify a location to filter to a specific site. Default is
\code{NULL}.}

\item{survey_year}{Specify a specific year. Default is \code{NULL}.}

\item{timezone}{Character. Specify a timezone for calculating sunset times.
Default is 'EST', do not change for Ontario data...}

\item{text_size}{Numeric: adjusts the size of text in the figure.}

\item{save_directory}{Character: if provided a .png image of the plot will be
saved in the folder specified. Defaults to \code{NULL}: no output saved.}

\item{date_breaks}{Date value: adjusts the formatting of the month labels on
the y-axis. See https://www.statmethods.net/input/dates.html for formatting.}
}
\value{
A plot as an object in the current envrionment, and a saved image if
 selected.
}
\description{
\code{first_observations_plot} creates a plot showing the first observation of
the specified species relative to local sunset for each site withn a project.
}
\concept{Activity Timing}
