% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_Plots.R
\name{monthly_activity_plot}
\alias{monthly_activity_plot}
\title{Monthly Activity Plot}
\usage{
monthly_activity_plot(
  species_night_site,
  monthly_active_nights,
  exclude_species = NULL,
  species_codes = c("Epfu", "Labo", "Laci", "Lano", "Myle", "Mylu", "Myse", "Mysp",
    "Mysp_all", "Pesu"),
  species_names = c("Big Brown Bat", "Eastern Red Bat", "Hoary Bat", "Silver-haired Bat",
    "Eastern Small-footed Myotis", "Little Brown Myotis", "Northern Myotis",
    "Unidentified Myotis", "Combined Myotis", "Tri-colored Bat")
)
}
\arguments{
\item{gaps}{Object: a \code{gaps_projectname} data frame genereted by the
\code{log_file_parser} function. Defaults to \code{null}.}

\item{project_name}{Character, a brief and relevant reference phrase that will
be used to name the text file. Ideally mathcing project name used in other
functions.}

\item{monitoring_start}{Character, the date the monitoring began (e.g.
"2019-01-01")}

\item{monitoring_end}{Character, the date the monitoring ceased (e.g.
"2019-01-01")}

\item{survey_year}{Character: for projects with data from multiple years, to
select year}

\item{save.directory}{Character: if provided a .png image of the plot will be
saved in the folder specified. Defaults to \code{NULL}: no output saved.}

\item{title}{Logical vector. Defaults \code{TRUE}, if \code{FALSE} title is
omitted}

\item{width}{Number. The width in cm of the plot if saved to file. Default is
25 cm.}

\item{height}{Number. The height of the plot if saved to file. Default is 20
cm.}

\item{text_size}{Numeric: adjusts the size of text in the figure.}

\item{date_breaks}{Date value: adjusts the formatting of the month labels on
the y-axis. See https://www.statmethods.net/input/dates.html for formatting.}
}
\value{
A plot as an object in the current envrionment, and a saved image if
 selected.
}
\description{
\code{monthly_actiivty_plot} creates a plot showing average nightly activity
by month and species.
}
\examples{
\dontrun{
species_daily_site_plot(species_night_site_projectname, "Project Name", "Mylu", "2019-01-01", "2019-12-31")
}
}
\seealso{
Other Basic Plots: 
\code{\link{monitoring_effort_plot}()},
\code{\link{multi_species_daily_site_plot}()},
\code{\link{old_species_daily_site_plot}()},
\code{\link{species_daily_site_plot}()},
\code{\link{species_site_aggregated_plot}()}
}
\concept{Basic Plots}
