% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_reports.R
\name{multi_site_short_report}
\alias{multi_site_short_report}
\title{Create Multi Site Report}
\usage{
multi_site_short_report(
  data_path,
  save_path,
  file_name,
  project_name,
  author,
  species,
  sites
)
}
\arguments{
\item{RData}{file}
}
\value{
A PDF report.
}
\description{
Create Multi Site Report
}
\examples{
\dontrun{
Insert Example!
}
}
