% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_tools.R
\name{one_site_short_report}
\alias{one_site_short_report}
\title{Create One Site Report}
\usage{
one_site_short_report(
  data_path,
  save_path,
  project_name,
  author,
  species,
  sites
)
}
\arguments{
\item{RData}{file}
}
\value{
A PDF report.
}
\description{
Create One Site Report
}
\examples{
\dontrun{
Insert Example!
}
}
