% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nightly_activity.R
\name{species_daily_site_plot}
\alias{species_daily_site_plot}
\title{Plot Activity for a Species at Each Site Within a Project}
\usage{
species_daily_site_plot(
  data_path,
  species,
  monitoring_start = NULL,
  monitoring_end = NULL,
  gaps = FALSE,
  save_directory = NULL,
  y_scale = "free_y",
  width = 15.9,
  height = 8.43,
  text_size = 10,
  date_label = "\%b",
  title = FALSE
)
}
\arguments{
\item{data_path}{Character. Path to an existing RData file.}

\item{species}{Character: the four letter species code for the species you
wish to plot. E.g. "Epfu".}

\item{monitoring_start}{Character, the date the monitoring began (e.g.
"2019-01-01"), leave as NULL to use the earliest data point as the
monitoring start.}

\item{monitoring_end}{Character, the date the monitoring ceased (e.g.
"2019-01-01"), leave as NULL to use the latest data point as the monitoring
end}

\item{gaps}{Boolean. Set true to add grey blocks to the plot to indicate
periods in which the recorder was not active (requires log files to be
loaded first using \code{import_logs}. Default is \code{FALSE}.)}

\item{y_scale}{Character. Determines whether scales on the y-axis are matched
or free. Defaults to "free_y" for unmatched axis, set to "fixed" for matched
axis.}

\item{width}{Number. The width in cm of the plot if saved to file. Default is
25 cm.}

\item{height}{Number. The height of the plot if saved to file. Default is 20
cm.}

\item{text_size}{Numeric: adjusts the size of text in the figure.}

\item{date_label}{Date value: adjusts the formatting of the month labels on
the y-axis. See https://www.statmethods.net/input/dates.html for formatting.}

\item{title}{Logical vector. If \code{FALSE} (default) title is omitted.}

\item{save.directory}{Character: if provided a .png image of the plot will be
saved in the folder specified. Defaults to \code{NULL}: no output saved.}
}
\value{
A plot as an object in the current environment, and a saved image if
 selected.
}
\description{
\code{species_daily_site_plot} creates a plot of nightly activity for a chosen
species at all sites within a project.
}
\examples{
\dontrun{
species_daily_site_plot(species_night_site_projectname, "Project Name", "Mylu", "2019-01-01", "2019-12-31")
}
}
\seealso{
Other Basic Plots: 
\code{\link{monitoring_effort_plot}()},
\code{\link{monthly_activity_plot}()},
\code{\link{multi_species_daily_site_plot}()},
\code{\link{old_species_daily_site_plot}()},
\code{\link{species_site_aggregated_plot}()}
}
\concept{Basic Plots}
